Extract CalculatorService interface
Move code from former CalculatorService into
ImperativeCalculatorService.
